+++ src/Integrations/AppIntegration.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace App\Integrations;
+
+use Rougin\Slytherin\Integration\Configuration;
+use Rougin\Slytherin\Container\ContainerInterface;
+
+/**
+ * Application Integration
+ *
+ * Bootstraps the additional definitions and integrate it to Slytherin.
+ *
+ * @package App
+ * @author  Rougin Royce Gutib <rougingutib@gmail.com>
+ */
+class AppIntegration implements \Rougin\Slytherin\Integration\IntegrationInterface
+{
+    /**
+     * Defines the specified integration.
+     *
+     * @param  \Rougin\Slytherin\Container\ContainerInterface $container
+     * @param  \Rougin\Slytherin\Integration\Configuration    $config
+     * @return \Rougin\Slytherin\Container\ContainerInterface
+     */
+    public function define(ContainerInterface $container, Configuration $config)
+    {
+        // Sets the default timezone
+        date_default_timezone_set($config->get('app.timezone'));
+
+        return $container;
+    }
+}