+++ src/Validators/UserValidator.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Skeleton\Validators;
+namespace App\Validators;
 
 /**
  * User Validator
  *
- * @package Skeleton
+ * @package App
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
  */
 class UserValidator extends AbstractValidator