+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-All Notable changes to `Skeleton` will be documented in this file
+All notable changes to `Skeleton` will be documented in this file.
 
 ## 0.1.0 - CURRENT_DATE