+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to `App` will be documented in this file.
 
-## 0.1.0 - CURRENT_DATE
+## 0.1.0 - Unreleased
 
 ### Added
 - `App` file and directory structure